64-Bit Patch #9 by [Timothy Wood]
This is the ninth patch submitted by Timothy Wood.  Fixes a few missed int/NSInteger issues

+++ Framework/BGHUDSegmentedCell.h
@@ -43,7 +43,7 @@
 
 @property (retain) NSString *themeKey;
 
--(void)drawInteriorForSegment:(int)segment withFrame:(NSRect)rect;
+-(void)drawInteriorForSegment:(NSInteger)segment withFrame:(NSRect)rect;
 -(BOOL)hasText;
 
 @end
